Move TestPlayers to its own file, and instead of having to change hundreds of files at once, remove a specific set (those that use more colors than red and blue.)
