+++ tests/Game.spec.ts
@@ -8,7 +8,8 @@ import * as constants from '../src/constants';
 import {Birds} from '../src/cards/base/Birds';
 import {WaterImportFromEuropa} from '../src/cards/base/WaterImportFromEuropa';
 import {Phase} from '../src/Phase';
-import {maxOutOceans, setCustomGameOptions, TestPlayers} from './TestingUtils';
+import {maxOutOceans, setCustomGameOptions} from './TestingUtils';
+import {TestPlayers} from './TestPlayers';
 import {SaturnSystems} from '../src/cards/corporation/SaturnSystems';
 import {Resources} from '../src/Resources';
 import {ISpace, SpaceId} from '../src/boards/ISpace';
