+++ tests/cards/moon/LunarSecurityStations.spec.ts
@@ -2,12 +2,13 @@ import {Game} from '../../../src/Game';
 import {IMoonData} from '../../../src/moon/IMoonData';
 import {MoonExpansion} from '../../../src/moon/MoonExpansion';
 import {Player} from '../../../src/Player';
-import {setCustomGameOptions, TestPlayers} from '../../TestingUtils';
+import {setCustomGameOptions} from '../../TestingUtils';
 import {LunarSecurityStations} from '../../../src/cards/moon/LunarSecurityStations';
 import {expect} from 'chai';
 import {OrOptions} from '../../../src/inputs/OrOptions';
 import {HiredRaiders} from '../../../src/cards/base/HiredRaiders';
 import {TileType} from '../../../src/TileType';
+import {TestPlayers} from '../../TestPlayers';
 
 const MOON_OPTIONS = setCustomGameOptions({moonExpansion: true});
