+++ tests/cards/promo/MonsInsurance.spec.ts
@@ -7,7 +7,7 @@ import {Game} from '../../../src/Game';
 import {OrOptions} from '../../../src/inputs/OrOptions';
 import {Player} from '../../../src/Player';
 import {Resources} from '../../../src/Resources';
-import {TestPlayers} from '../../TestingUtils';
+import {TestPlayers} from '../../TestPlayers';
 
 describe('MonsInsurance', function() {
   let card : MonsInsurance; let player : Player; let player2: Player; let player3: Player;
