+++ tests/cards/promo/SmallAsteroid.spec.ts
@@ -4,7 +4,7 @@ import {Game} from '../../../src/Game';
 import {OrOptions} from '../../../src/inputs/OrOptions';
 import {Player} from '../../../src/Player';
 import {Resources} from '../../../src/Resources';
-import {TestPlayers} from '../../TestingUtils';
+import {TestPlayers} from '../../TestPlayers';
 
 describe('SmallAsteroid', function() {
   let card : SmallAsteroid; let player : Player; let player2 : Player;
